+++ rpcclient.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "rpccall.h"
+#include <stdint.h>
 
 enum rpcclient_disconnect_reason{
    INITIATED,
@@ -25,7 +26,7 @@ struct rpcclient_fninfo{
 
 
 struct rpcclient* rpcclient_connect(char* host,int portno,char* key);
-int rpcclient_call(struct rpcclient* self,char* fn,enum rpctypes* rpctypes,char* flags, int rpctypes_len,void* fnret,...);
+int rpcclient_call(struct rpcclient* self,char* fn,enum rpctypes* rpctypes,char* flags, uint8_t rpctypes_len,void* fnret,...);
 void rpcclient_disconnect(struct rpcclient* self);
 struct rpcclient_fninfo* rpcclient_list_functions(struct rpcclient* self,size_t* fn_len);
 void rpcclient_fninfo_free(struct rpcclient_fninfo* in,size_t len);
